+++ src/features/browser.client/index.ts
@@ -2,4 +2,5 @@ export * from "./clipboard";
 export * from "./debounce";
 export * from "./geolocation";
 export * from "./haptics";
+export * from "./preferences-storage";
 export * from "./share";